kafka/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["lucykomarovskikh <lucy_komar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
pydantic = "^1.10.8"
elasticsearch = "^8.8.0"
uvicorn = "^0.22.0"
redis = "^4.5.5"
aioredis = "^2.0.1"
orjson = "^3.9.0"
loguru = "^0.7.0"
aiohttp = "^3.8.4"
confluent-kafka = "^2.2.0"
aiokafka = "^0.8.1"
python-schema-registry-client = "^2.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
